backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
SpeechRecognition==3.10.0
gTTS==2.3.2
pymongo==4.13.0
dnspython==2.8.0
